LayoutTests/svg/animations/animations-paused-page-non-visible-expected.txt

Tests that SVG animations are properly paused when the page becomes hidden.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.areSVGAnimationsPaused is false
* Page is now hidden
PASS document.hidden is true
PASS internals.areSVGAnimationsPaused is true
* Page is now visible
PASS document.hidden is false
PASS internals.areSVGAnimationsPaused is false
PASS successfullyParsed is true

TEST COMPLETE
